+++ tech1-framework-b2b-mongodb-security-jwt/src/main/java/io/tech1/framework/b2b/mongodb/security/jwt/configurations/ApplicationBaseSecurityJwt.java
@@ -53,6 +53,8 @@ import static org.springframework.http.HttpMethod.*;
         "io.tech1.framework.b2b.mongodb.security.jwt.utilities",
         "io.tech1.framework.b2b.mongodb.security.jwt.validators",
         // -------------------------------------------------------------------------------------------------------------
+        "io.tech1.framework.utilities.geo"
+        // -------------------------------------------------------------------------------------------------------------
 })
 @EnableWebSecurity
 @Import({