+++ tech1-framework-utilities/src/main/java/io/tech1/framework/utilities/geo/functions/mindmax/impl/MindMaxGeoLocationUtilityImpl.java
@@ -1,10 +1,10 @@
-package io.tech1.framework.b2b.mongodb.security.jwt.utilities.impl;
+package io.tech1.framework.utilities.geo.functions.mindmax.impl;
 
 import com.maxmind.geoip2.DatabaseReader;
 import com.maxmind.geoip2.exception.GeoIp2Exception;
-import io.tech1.framework.b2b.mongodb.security.jwt.utilities.GeoUtility;
 import io.tech1.framework.domain.geo.GeoLocation;
 import io.tech1.framework.domain.http.requests.IPAddress;
+import io.tech1.framework.utilities.geo.functions.mindmax.MindMaxGeoLocationUtility;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.ResourceLoader;
@@ -20,13 +20,13 @@ import static io.tech1.framework.domain.enums.Status.SUCCESS;
 
 @Slf4j
 @Component
-public class GeoUtilityImpl implements GeoUtility {
+public class MindMaxGeoLocationUtilityImpl implements MindMaxGeoLocationUtility {
     private static final String GEO_DATABASE_NAME = "GeoLite2-City.mmdb";
 
     private final DatabaseReader databaseReader;
 
     @Autowired
-    public GeoUtilityImpl(ResourceLoader resourceLoader) {
+    public MindMaxGeoLocationUtilityImpl(ResourceLoader resourceLoader) {
         try {
             var resource = resourceLoader.getResource("classpath:" + GEO_DATABASE_NAME);
             var inputStream = resource.getInputStream();