+++ tech1-framework-utilities/src/test/java/io/tech1/framework/utilities/environment/base/BaseEnvironmentUtilityTest.java
@@ -86,6 +86,7 @@ public class BaseEnvironmentUtilityTest {
 
         // Assert
         verify(this.environment).getActiveProfiles();
+        assertThat(throwable.getClass()).isEqualTo(IllegalArgumentException.class);
         assertThat(throwable.getMessage()).isEqualTo("Base Environment Utility contains ONLY one active profile");
     }