+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 
 ### v1.6
 - Add "countryCode" -> GeoLocation
+- Add geo location facade utility (ipapi + mindmax as failsafe)
 
 ### v1.5
 - Migrate EnumUtility methods, TuplePercentage(long, long)